refactor: 기본적인 컴포넌트 base 폴더 이동
