+++ client/src/components/index.tsx
@@ -1,5 +1,5 @@
 export { default as Advertisement } from '@/components/common/Advertisement';
-export { default as Layout } from '@/components/common/Layout';
+export { default as Layout } from '@/components/base/Layout';
 export { default as MenuView } from '@/components/common/MenuView';
 export { default as Navigator } from '@/components/common/Navigator';
 export {
@@ -12,7 +12,7 @@ export { default as QuillEditBox } from '@/components/common/QuillEditBox';
 export { default as Divider } from '@/components/common/Divider';
 export { default as Avatar } from '@/components/common/Avatar';
 export { default as LoadingIcon } from '@/components/common/LoadingIcon';
-export { default as Icon } from '@/components/common/Icon';
+export { default as Icon } from '@/components/base/Icon';
 export { default as Modal } from '@/components/common/Modal';
 export { default as AuthComponent } from '@/components/common/AuthComponent';
 export { default as EditProfile } from '@/components/common/EditProfile';
